\name{EstDemand2-package}
\alias{EstDemand2-package}
\alias{EstDemand2}
\docType{package}
\title{
\packageTitle{EstDemand2}
}
\description{
\packageDescription{EstDemand2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EstDemand2}
\packageIndices{EstDemand2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EstDemand2}

Maintainer: \packageMaintainer{EstDemand2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
